requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
django-service-objects==0.7.0
djangorestframework-simplejwt==4.8.0